pyproject.toml
[project]
name = "emqx-knowledge-base"
version = "0.1.0"
description = "EMQX Knowledge Base with Slack integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.21.0",
    "openai>=1.3.0",
    "psycopg[binary,pool]>=3.1.12",
    "pgvector>=0.2.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "requests>=2.31.0",
    "uvicorn[standard]>=0.34.0",
    "fastapi>=0.115.11",
    "python-multipart>=0.0.20",
    "llama-index-core>=0.12.24.post1",
    "llama-index-llms-openai>=0.3.25",
    "sse-starlette>=2.0.0",
    "starlette>=0.46.1",
    "itsdangerous>=2.2.0",
    "PyJWT>=2.8.0",
    "llama-index-embeddings-openai>=0.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
